backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
celery==5.3.0
redis==4.5.0
django-environ==0.8.1
djangorestframework-simplejwt==5.2.2
django-allauth==0.51.0
sendgrid==6.9.1